Issue-258: Add SecureBatchReader Example (#257)
This change is for demonstrating batch read with tls configuration. This particular case was found to be tricky on a Kubernetes deployment because it involved setting up the PravegaCluster with K8s services and tls which introduces a relay. On the other hand, this creates a standalone program which makes it easier to use with the already popular secureWriter and secureReader.

Signed-off-by: Ravi Rajamani <ravi.rajamani@emc.com>
Signed-off-by: Fan, Yang <fan.yang5@emc.com>